Brewverse complete
Added additional components for modularity to be used in the Brewverse
component. Finalized camera controls so that the map can be moved
around while preserving the zoom animation. Added limit controls to the
map so that it can only move so far in each direction. Implemented scene
transitions between buildings and their interiors.