Fix brewverse resize
The Brewverse now properly scales to the window size, while keeping the
correct aspect ratio of the base layer so that everything on the map
remains in view. Dragging the map is still fully functional but the
offset is set to 0 since the map is in full view.

Also created a Connect component that can be placed on the Nav to
connect a user's MetaMask wallet to the site.