Globalized the ether.js provider state
Created a wrapper component that provides access to the current ether.js
provider at the component level. The provider can now be accessed and be
set in different components without having to recreate the provider each
time.

Updated the connect button so that once the wallet is connected,
clicking it again disconnects it.