Implemented wallet authentication
When a user connects their wallet, a nonce is generated on the server
and stored in a gun datastore. On the client this nonce is signed by the
user, with the signature being used to verify whether the address of the
wallet matches that of the sigee.

Each time the user connects a new nonce is generated and stored on the
backend. Will likely need to refactor the datastore to make sure it is
updating the data properly.