Refactored wallet connection
In authenticate.js, instead of saving the nonce as an object in the gun
database, it saves it directly to the address. Added error checking in
verify.js in the case that the nonce corresponding to the address can't
be found for whatever reason.

In the Connect component, added error handling for when a user cancels
the first metamask connection for an account.

Updated the Keg Plebs brewery textures with the new transparent ones.