Updated navbar user interaction
The hamburger menu icon is now transition to and from an 'X' when clicked.
All section titles and icons change color to pink when hovering, except
for the Etherscan/Opensea images.

Might want to change the Etherscan/Opensea images to svgr to make it so
we can adjust their color to make them match